+++ CHANGELOG.md
@@ -9,7 +9,9 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
-- Changed the condition type of `HAS` arguments in `@whereConditions` to always use `String` type https://github.com/nuwave/lighthouse/pull/2725
+### Fixed
+
+- Use `String` type for `HAS.condition.column` in `@whereConditions` instead of incorrectly inheriting the parent's column enum type https://github.com/nuwave/lighthouse/pull/2725
 
 ## v6.63.1